Fix: auto correlation matrix is indefinite
The auto correlation matrix as computed by BaseFunction.autoCorrPSF is sometimes indefinite. This was a result of the linear interpolation of the auto correlation function. In this commit pre-computation of auto correlation function is removed. On every function call the PSF is convolved by itself and the result is transformed into a toeplitz matrix. This way the semi-definiteness of the auto correlation matrix is guaranteed.