Fix: dereferencing null pointer
MeshFitterState::hiddenState_ is usially never nullptr when used properly. Hoewever, in the C bindings there is no extra object class for MeshFitter::Result and MeshFitter::State is used instead. When now a MeshFitter::Result is used, e.g. as returned from MeshFitter::fit(), to initialize a State object, hiddenState_ is null, obviously.
Added a simple check for nullptr to avoid dereferencing a null pointer in the constructor.
Fixes #70.